Refactored Directory Structure
