+++ services/api/src/routes/__openapi__/categories.json
