+++ services/api/src/routes/__openapi__/shops.json
