+++ services/api/src/routes/__openapi__/uploads.json
