+++ services/api/src/routes/__tests__/fixtures/logo.png
