+++ services/api/src/routes/__tests__/invites.js
@@ -1,4 +1,4 @@
-const { setupDb, teardownDb, request, createUser } = require('../../test-helpers');
+const { setupDb, teardownDb, request, createUser } = require('../../utils/testing');
 const { Invite, User } = require('../../models');
 
 beforeAll(async () => {
