+++ services/api/src/utils/__tests__/fixtures/custom-model.json
