+++ services/api/src/utils/__tests__/fixtures/special-category.json
