+++ services/api/src/utils/middleware/__tests__/validate.js
@@ -1,11 +1,11 @@
 const Joi = require('@hapi/joi');
 const validate = require('../validate');
-const { context } = require('../../test-helpers');
+const { context } = require('../../testing');
 
 describe('validate', () => {
   it("should throw an error when the request doesn't contain a key as specified in Joi schema", async () => {
     const middleware = validate({
-      body: Joi.object({}) // Does not exist in given context
+      body: Joi.object({}), // Does not exist in given context
     });
     const ctx = context();
 
@@ -18,7 +18,7 @@ describe('validate', () => {
   it('should reject a request with invalid params', async () => {
     // Require test param, but don't provide it in ctx object:
     const middleware = validate({
-      body: Joi.object().keys({ test: Joi.number().required() })
+      body: Joi.object().keys({ test: Joi.number().required() }),
     });
 
     const ctx = context();
@@ -29,7 +29,7 @@ describe('validate', () => {
 
   it('should accept a valid request', async () => {
     const middleware = validate({
-      body: Joi.object().keys({ test: Joi.string().required() })
+      body: Joi.object().keys({ test: Joi.string().required() }),
     });
     const ctx = context({ url: '/' });
     ctx.request.body = { test: 'something' };
@@ -41,7 +41,7 @@ describe('validate', () => {
 
   it('should support the light syntax', async () => {
     const middleware = validate({
-      body: Joi.object({ test: Joi.string().required() })
+      body: Joi.object({ test: Joi.string().required() }),
     });
     const ctx = context({ url: '/' });
     ctx.request.body = { test: 'something' };
@@ -53,11 +53,11 @@ describe('validate', () => {
 
   it('should do type conversion for query', async () => {
     const middleware = validate({
-      query: Joi.object({ convertToNumber: Joi.number().required() })
+      query: Joi.object({ convertToNumber: Joi.number().required() }),
     });
     const ctx = context({ url: '/' });
     ctx.request.query = {
-      convertToNumber: '1234'
+      convertToNumber: '1234',
     };
 
     await middleware(ctx, () => {
@@ -67,13 +67,13 @@ describe('validate', () => {
 
   it('should not allow attributes that are not defined', async () => {
     const middleware = validate({
-      query: Joi.object({ somethingExisting: Joi.string() })
+      query: Joi.object({ somethingExisting: Joi.string() }),
     });
 
     const ctx = context({ url: '/' });
     ctx.request.query = {
       somethingExisting: 'yes',
-      shouldBeRemoved: 'should be been removed from request'
+      shouldBeRemoved: 'should be been removed from request',
     };
 
     await expect(middleware(ctx)).rejects.toHaveProperty('status', 400);
